core: classify a pre-voice DMR LC under the forced ALG ID, and consolidate the tg-key-map seams
Trunk tuning zeroes payload_algid on every voice-channel tune and
--dmr-force-algid is only installed on the first voice frame, so on a forced-ALGID
system every trunked call's first voice LC arrived with no ALG ID. The LC path then
classified it as "ALG unknown" -- against whatever key the slot last carried (R is 0
on a fresh session) -- instead of against the key the forced ALG and any
--dmr-tg-key-csv row were about to select. The very call the map exists to decrypt
could publish ENCRYPTED and, under --enc-lockout, be locked out and dropped before its
first voice frame.

dsd_dmr_classify_algid() is the read-only twin of dsd_dmr_apply_forced_algid(): both
derive the forced value from one helper, so the fallback rule (#352: OTA always wins,
only an encrypted slot with no ALG ID borrows it) cannot drift between them.
dmr_flco_publish_crypto() and dmr_flco_slot_can_decrypt() classify under it; the
published .algid and the slot stay OTA, and the 0xFF key-id sentinel is still written
only by the voice path. Pinned in DMR_FLCO_PRIVACY_MODES (red without the change) and
CORE_DMR_VOICE_ALG_GATE.

Review follow-ups on the same seams:

- The <= 0xFF key-id width guard now lives in keyring_dmr_effective_kid() (int in, int
  out); the seven hand-copied guards at the voice, LC, PI, PDU and three sdrtrunk
  replay sites are gone. Pinned in CORE_DMR_TG_KEY_MAP.
- The talkgroup->key-id policy and its operator notices move from keyring.c to
  keyring_dmr_tg_map.c; keyring.c is back to key material only.
- dsd_dmr_voice_kid_can_decrypt() takes a const dsd_dmr_key_material* instead of the
  same three fields as loose parameters.
- csvDmrTgKeyImport() and csvVertexKsImport() share one csv_mapping_import_rows()
  skeleton (open, header skip, trim, parse-row, refuse-on-first-error).
- Angle-bracket project includes in keyring.c and the two new tests; the block-crypto
  key-map test picks up dmr_block_crypto.h through the include dir like its sibling.
- docs/cli.md and docs/csv-formats.md describe the pre-voice classification rule.

+++ include/dsd-neo/core/keyring.h
@@ -40,8 +40,8 @@ void keyring_dmr_tg_map_reset(dsd_state* state);
  * Use this on the voice path, where a call epoch exists to latch the notice against. Other
  * consumers want keyring_dmr_kid_for_call(), which is silent.
  */
-uint8_t keyring_dmr_slot_kid_for_call(dsd_state* state, int slot, const dsd_call_snapshot* call,
-                                      dsd_key_material_need need, uint8_t signaled_kid);
+int keyring_dmr_slot_kid_for_call(dsd_state* state, int slot, const dsd_call_snapshot* call, dsd_key_material_need need,
+                                  int signaled_kid);
 
 /**
  * Report the imported key material behind a key ID without activating it.
@@ -103,12 +103,15 @@ int keyring_kid_satisfies_need(const dsd_state* state, int key_id, dsd_key_mater
  * 24-bit space, so an unchecked match keys a unit call off a colliding row.
  *
  * Returns `signaled_kid` unchanged whenever the map does not apply, so callers activate one
- * key ID rather than maintaining a map path and a signaled path in parallel.
+ * key ID rather than maintaining a map path and a signaled path in parallel. That includes a
+ * `signaled_kid` outside 0..0xFF: DMR key IDs are byte-wide, but `payload_keyid` is shared
+ * across protocols and P25 stores a full 16-bit KID there, so the resolver owns the width guard
+ * and every caller passes the value it holds without narrowing it first.
  *
  * @param out_mapped set to 1 when a map row was applied, 0 otherwise (may be NULL)
  */
-uint8_t keyring_dmr_effective_kid(const dsd_state* state, uint32_t target, int target_is_group,
-                                  dsd_key_material_need need, uint8_t signaled_kid, int* out_mapped);
+int keyring_dmr_effective_kid(const dsd_state* state, uint32_t target, int target_is_group, dsd_key_material_need need,
+                              int signaled_kid, int* out_mapped);
 
 /** Activate imported key material for a slot using an explicit key ID. */
 void keyring_activate_slot_with_kid(dsd_state* state, int slot, int key_id);
@@ -119,8 +122,8 @@ void keyring_activate_slot_with_kid(dsd_state* state, int slot, int key_id);
  * Applies only to an active DMR group-voice call with a usable talkgroup; returns
  * `signaled_kid` for anything else. Announces nothing -- see keyring_dmr_slot_kid_for_call().
  */
-uint8_t keyring_dmr_kid_for_call(const dsd_state* state, const dsd_call_snapshot* call, dsd_key_material_need need,
-                                 uint8_t signaled_kid, int* out_mapped);
+int keyring_dmr_kid_for_call(const dsd_state* state, const dsd_call_snapshot* call, dsd_key_material_need need,
+                             int signaled_kid, int* out_mapped);
 
 #ifdef __cplusplus
 }
